bridge/requirements.txt
python-telegram-bot==20.8
paho-mqtt==1.6.1
python-dotenv